+++ app/models/visual_event.py
@@ -6,7 +6,7 @@ import uuid
 from datetime import UTC, datetime
 from typing import TYPE_CHECKING
 
-from sqlalchemy import DateTime, Float, ForeignKey, Integer, String, Text
+from sqlalchemy import BigInteger, DateTime, Float, ForeignKey, String, Text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db.base import Base
@@ -33,7 +33,7 @@ class VisualEvent(Base):
         nullable=False,
         index=True,
     )
-    timestamp_ms: Mapped[int] = mapped_column(Integer, nullable=False)
+    timestamp_ms: Mapped[int] = mapped_column(BigInteger, nullable=False)
     source: Mapped[str] = mapped_column(String(16), nullable=False)
     ocr_text: Mapped[str] = mapped_column(Text, nullable=False)
     ocr_confidence: Mapped[float] = mapped_column(Float, nullable=False)
